Add glaze/forward.hpp for lightweight customization-point forward declarations
Users who want to specialize glz::meta, glz::to<Format, T>, glz::from<Format, T>,
or construct glz::custom_t in widely-included user headers can now include only
<glaze/forward.hpp>, which has no Glaze dependencies and pulls in <cstdint> only.
This lets them defer the full <glaze/glaze.hpp> include to the translation units
that actually call read/write.

New: include/glaze/forward.hpp — format id constants, primary templates for
meta/to/from/to_partial/skip_value/parse/serialize/serialize_partial, the
custom_t aggregate, and a namespace-scoped `using uint32_t = std::uint32_t;`
so the format-tag type has a home under glz::.

core/opts.hpp, core/meta.hpp, core/wrappers.hpp now include forward.hpp and
delete their duplicate declarations. Behavior is unchanged.

Addresses #2525.
