+++ include/glaze/core/opts.hpp
@@ -9,35 +9,12 @@
 #include "glaze/core/context.hpp"
 #include "glaze/core/optimization_level.hpp"
 #include "glaze/core/traits.hpp"
+#include "glaze/forward.hpp"
 #include "glaze/util/inline.hpp"
 #include "glaze/util/type_traits.hpp"
 
 namespace glz
 {
-   // Formats
-   // Built in formats must be less than 65536
-   // User defined formats can be 65536 to 4294967296
-   inline constexpr uint32_t INVALID = 0;
-   inline constexpr uint32_t BEVE = 1;
-   inline constexpr uint32_t CBOR = 2; // RFC 8949 - Concise Binary Object Representation
-   inline constexpr uint32_t JSONB = 3; // SQLite JSONB binary JSON format - https://sqlite.org/jsonb.html
-   inline constexpr uint32_t BSON = 4; // MongoDB BSON 1.1 - https://bsonspec.org/spec.html
-   inline constexpr uint32_t JSON = 10;
-   inline constexpr uint32_t JSON_PTR = 20;
-   inline constexpr uint32_t MSGPACK = 30;
-   inline constexpr uint32_t NDJSON = 100; // new line delimited JSON
-   inline constexpr uint32_t TOML = 400;
-   inline constexpr uint32_t YAML = 450;
-   inline constexpr uint32_t STENCIL = 500;
-   inline constexpr uint32_t MUSTACHE = 501;
-   inline constexpr uint32_t CSV = 10000;
-   inline constexpr uint32_t EETF = 20000;
-
-   // Protocol formats
-   inline constexpr uint32_t REPE = 30000;
-   inline constexpr uint32_t REST = 30100;
-   inline constexpr uint32_t JSONRPC = 30200;
-
    // layout
    inline constexpr uint8_t rowwise = 0;
    inline constexpr uint8_t colwise = 1;
@@ -197,7 +174,7 @@ namespace glz
    // Format string for floating-point output using std::format (C++23)
    // Must be a static constexpr member to be usable as a template parameter
    // Example: "{:.2f}" for 2 decimal places, "{:.6g}" for 6 significant digits, "{:.0f}" for integers
-   // When specified, uses std::format_to instead of Dragonbox for float/double serialization
+   // When specified, uses std::uint32_to instead of Dragonbox for float/double serialization
 
    // ---
    // bool skip_read_constraint = false;
@@ -1479,36 +1456,15 @@ namespace glz
 
 namespace glz
 {
-   template <uint32_t Format = INVALID, class T = void>
-   struct to;
-
-   template <uint32_t Format = INVALID, class T = void>
-   struct from;
-
-   // Note: specified and is_specified are defined in glaze/core/traits.hpp
-
-   template <uint32_t Format = INVALID, class T = void>
-   struct to_partial;
-
-   template <uint32_t Format = INVALID>
-   struct skip_value;
+   // Note: primary templates for glz::to, glz::from, glz::to_partial, glz::skip_value,
+   // glz::parse, glz::serialize, glz::serialize_partial live in glaze/forward.hpp.
+   // Note: specified and is_specified are defined in glaze/core/traits.hpp.
 
    template <class T, uint32_t Format>
    concept write_supported = requires { to<Format, std::remove_cvref_t<T>>{}; };
 
    template <class T, uint32_t Format>
    concept read_supported = requires { from<Format, std::remove_cvref_t<T>>{}; };
-
-   // These templates save typing by determining the core type used to select the proper to/from specialization
-   // Long term I would like to remove these detail indirections.
-
-   template <uint32_t Format>
-   struct parse
-   {};
-
-   template <uint32_t Format>
-   struct serialize
-   {};
 
    // Default context type for a format.
    // Formats that need a richer context (e.g. YAML) specialize this.
@@ -1520,10 +1476,6 @@ namespace glz
 
    template <uint32_t Format>
    using format_context_t = typename format_context<Format>::type;
-
-   template <uint32_t Format>
-   struct serialize_partial
-   {};
 
    // Preset options for size-optimized builds (embedded systems)
    struct opts_size : opts
